refactor: 쿼리를 활용한 구현

+++ src/main/kotlin/com/group/libraryapp/controller/book/BookController.kt
@@ -33,7 +33,7 @@ class BookController(
     }
 
     @GetMapping("/book/loan")
-    fun countLoanedBook(): Int {
+    fun countLoanedBook(): Long {
         return bookService.countLoanedBook()
     }
